requirements.txt
hamolicious-vector-class==0.2.10
pygame==2.0.1
